reverts the changes made to the handleDimensionErr func - not in scope of the current 1351 ticket
