+++ api/common/errors.go
@@ -6,51 +6,8 @@ import (
 	"strings"
 
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
-	dprequest "github.com/ONSdigital/dp-net/v2/request"
 	"github.com/ONSdigital/log.go/v2/log"
-	"github.com/pkg/errors"
 )
-
-func HandlePatchReqErr(ctx context.Context, w http.ResponseWriter, err error, data log.Data) {
-	if data == nil {
-		data = log.Data{}
-	}
-
-	var status int
-
-	// Switch by error type
-	switch err.(type) {
-	case errs.ErrInvalidPatch:
-		status = http.StatusBadRequest
-	default:
-		// Switch by error message
-		switch {
-		case errs.NotFoundMap[err]:
-			status = http.StatusNotFound
-		case errs.BadRequestMap[err]:
-			status = http.StatusBadRequest
-		case errs.ConflictRequestMap[err]:
-			status = http.StatusConflict
-		default:
-			status = http.StatusInternalServerError
-			err = errors.WithMessage(err, "internal error")
-		}
-	}
-
-	data["response_status"] = status
-	logError(ctx, err, data)
-	http.Error(w, err.Error(), status)
-}
-
-func logError(ctx context.Context, err error, data log.Data) {
-	if user := dprequest.User(ctx); user != "" {
-		data[reqUser] = user
-	}
-	if caller := dprequest.Caller(ctx); caller != "" {
-		data[reqCaller] = caller
-	}
-	log.Error(ctx, "unsuccessful request", err, data)
-}
 
 func HandleDatasetAPIErr(ctx context.Context, err error, w http.ResponseWriter, data log.Data) {
 	if data == nil {
