+++ dimension/helpers.go
@@ -2,6 +2,7 @@ package dimension
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"io"
 	"io/ioutil"
@@ -10,7 +11,14 @@ import (
 
 	errs "github.com/ONSdigital/dp-dataset-api/apierrors"
 	"github.com/ONSdigital/dp-dataset-api/models"
+	dprequest "github.com/ONSdigital/dp-net/v2/request"
+	"github.com/ONSdigital/log.go/v2/log"
 	"github.com/pkg/errors"
+)
+
+const (
+	reqUser   = "req_user"
+	reqCaller = "req_caller"
 )
 
 // Regex const for patch paths
@@ -87,6 +95,47 @@ func unmarshalDimensionCache(reader io.Reader) (*models.CachedDimensionOption, e
 	}
 
 	return &option, nil
+}
+
+func handleDimensionErr(ctx context.Context, w http.ResponseWriter, err error, data log.Data) {
+	if data == nil {
+		data = log.Data{}
+	}
+
+	var status int
+
+	// Switch by error type
+	switch err.(type) {
+	case errs.ErrInvalidPatch:
+		status = http.StatusBadRequest
+	default:
+		// Switch by error message
+		switch {
+		case errs.NotFoundMap[err]:
+			status = http.StatusNotFound
+		case errs.BadRequestMap[err]:
+			status = http.StatusBadRequest
+		case errs.ConflictRequestMap[err]:
+			status = http.StatusConflict
+		default:
+			status = http.StatusInternalServerError
+			err = errors.WithMessage(err, "internal error")
+		}
+	}
+
+	data["response_status"] = status
+	logError(ctx, err, data)
+	http.Error(w, err.Error(), status)
+}
+
+func logError(ctx context.Context, err error, data log.Data) {
+	if user := dprequest.User(ctx); user != "" {
+		data[reqUser] = user
+	}
+	if caller := dprequest.Caller(ctx); caller != "" {
+		data[reqCaller] = caller
+	}
+	log.Error(ctx, "unsuccessful request", err, data)
 }
 
 func setJSONPatchContentType(w http.ResponseWriter) {
